+++ .babelrc
@@ -1,6 +1,6 @@
 {
     "presets": [
-        ["@babel/preset-env", { "useBuiltIns": "usage", "corejs": 3 }],
+        ["@babel/preset-env", { "targets": { "node": "current" } }],
         "@babel/preset-typescript"
     ]
 }
