+++ .vscode/settings.json
@@ -2,7 +2,8 @@
     "cSpell.words": [
         "Fieldguide",
         "Hasura",
-        "Postgres"
+        "Postgres",
+        "strava"
     ],
     "typescript.tsdk": "node_modules/typescript/lib"
 }
