+++ assets/js/stream/types.ts
@@ -1,0 +1,1 @@
+export type RandomIntegerFactory = (min: number, max: number) => number;
