+++ package.json
@@ -6,11 +6,11 @@
     "private": true,
     "scripts": {
         "start": "hugo serve",
-        "test": "karma start --single-run",
+        "test": "jest",
         "posttest": "npm run lint-js && npm run prettier && npm run lint-sass",
         "predeploy": "hugo --minify",
         "deploy": "gulp deploy",
-        "lint-js": "eslint . --ext .js,.ts",
+        "lint-js": "tsc --noEmit && eslint . --ext .js,.ts",
         "lint-sass": "stylelint \"assets/**/*.scss\"",
         "prettier": "prettier --check ."
     },
@@ -28,29 +28,25 @@
         "@babel/core": "^7.16.0",
         "@babel/preset-env": "^7.16.4",
         "@babel/preset-typescript": "^7.16.0",
-        "@types/jasmine": "^3.10.2",
+        "@types/jest": "^27.0.3",
         "@types/jquery": "^3.5.9",
         "@types/moment": "^2.13.0",
         "@typescript-eslint/eslint-plugin": "^4.33.0",
         "@typescript-eslint/parser": "^4.33.0",
+        "babel-jest": "^27.4.5",
         "babel-loader": "^8.2.3",
         "config": "^3.3.6",
         "eslint": "^7.32.0",
         "eslint-config-prettier": "^8.3.0",
         "fancy-log": "^1.3.3",
         "gulp": "^4.0.2",
-        "jasmine-core": "^3.10.1",
-        "karma": "^6.3.9",
-        "karma-chrome-launcher": "^3.1.0",
-        "karma-jasmine": "^4.0",
-        "karma-webpack": "^5.0.0",
+        "jest": "^27.4.5",
         "prettier": "2.5.1",
         "stylelint": "^13.13.1",
         "stylelint-config-prettier": "^9.0.3",
         "stylelint-config-sass-guidelines": "^8.0.0",
         "typescript": "^4.5.2",
-        "vinyl-ftp": "^0.6.1",
-        "webpack": "^5.64.4"
+        "vinyl-ftp": "^0.6.1"
     },
     "engines": {
         "node": ">=10"
