+++ tsconfig.json
@@ -6,7 +6,8 @@
         "noEmit": true,
         "strict": true,
         "target": "es2015",
-        "types": ["jasmine", "jquery", "moment"]
+        "types": ["jest", "jquery", "moment"],
+        "allowSyntheticDefaultImports": true
     },
     "exclude": ["node_modules"]
 }
